feat: refactored version
