+++ src/hooks/useOffline.js
@@ -1,0 +1,22 @@
+import { useEffect, useState } from 'react'
+
+export function useOffline() {
+  const [offline, setOffline] = useState(false);
+
+  useEffect(() => {
+    function handleOffline() {
+      setOffline(true);
+    }
+    function handleOnline() {
+      setOffline(false);
+    }
+    window.addEventListener("offline", handleOffline);
+    window.addEventListener("online", handleOnline);
+    return () => {
+      window.removeEventListener("offline", handleOffline);
+      window.removeEventListener("online", handleOnline);
+    };
+  }, []);
+
+  return offline
+}
